# target/classes/srs-openapi.yaml
openapi: "3.0.3"
info:
  title: "srs API"
  description: "srs API"
  version: "1.0.0"
servers:
  - url: "https://srs"
paths:
...
  /prisons/add:
    post:
      summary: "POST prisons/add"
      operationId: "addPrison"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrisonDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /prisons/delete/{id}:
    delete:
      summary: "DELETE prisons/delete/{id}"
      operationId: "deletePrison"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /prisons/filter:
    get:
      summary: "GET prisons/filter"
      operationId: "getPrisonsByFilter"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrisonFilter"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/PrisonDTO"
  /prisons/update:
    put:
      summary: "PUT prisons/update"
      operationId: "updatePrison"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrisonDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
...
  /prisoners/add:
    post:
      summary: "POST prisoners/add"
      operationId: "addPrisoner"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrisonerDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /prisoners/delete/{id}:
    delete:
      summary: "DELETE prisoners/delete/{id}"
      operationId: "deletePrisoner"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /prisoners/filter:
    get:
      summary: "GET prisoners/filter"
      operationId: "getPrisonersByFilter"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrisonerFilter"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/PrisonerDTO"
  /prisoners/update:
    put:
      summary: "PUT prisoners/update"
      operationId: "updatePrisoner"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrisonerDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
...
  /reports/add:
    post:
      summary: "POST reports/add"
      operationId: "addReport"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /reports/delete/{id}:
    delete:
      summary: "DELETE reports/delete/{id}"
      operationId: "deleteReport"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /reports/filter:
    get:
      summary: "GET reports/filter"
      operationId: "getReportsByFilter"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportFilter"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/ReportDTO"
  /reports/update:
    put:
      summary: "PUT reports/update"
      operationId: "updateReport"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
...
  /roles/add:
    post:
      summary: "POST roles/add"
      operationId: "addRole"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /roles/delete/{id}:
    delete:
      summary: "DELETE roles/delete/{id}"
      operationId: "deleteRole"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /roles/filter:
    get:
      summary: "GET roles/filter"
      operationId: "getRolesByFilter"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleFilter"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/RoleDTO"
  /roles/update:
    put:
      summary: "PUT roles/update"
      operationId: "updateRole"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
...
components:
  schemas:
...
    OperationStatus:
      type: "object"
      properties:
        targetClassName:
          type: "string"
        recordId:
          type: "integer"
          format: "int32"
        operationName:
          type: "string"
        success:
          type: "boolean"
        msgList:
          type: "array"
          items:
            type: "string"
...
    PrisonDTO:
      type: "object"
      properties: { }
    PrisonFilter:
      type: "object"
      properties: { }
    PrisonerDTO:
      type: "object"
      properties:
        prisonerId:
          type: "integer"
          format: "int32"
        name:
          type: "string"
        surname:
          type: "string"
        pesel:
          type: "string"
        kartaPobytuId:
          type: "string"
        passportId:
          type: "string"
    PrisonerFilter:
      type: "object"
      properties: { }
    ReportDTO:
      type: "object"
      properties:
        reportId:
          type: "integer"
          format: "int32"
        prisonId:
          type: "integer"
          format: "int32"
        priorityId:
          type: "integer"
          format: "int32"
        shiftId:
          type: "integer"
          format: "int32"
        segment:
          type: "string"
        blok:
          type: "string"
        description:
          type: "string"
    ReportFilter:
      type: "object"
      properties: { }
    RoleDTO:
      type: "object"
      properties:
        roleId:
          type: "integer"
          format: "int32"
        roleCode:
          type: "string"
        name:
          type: "string"
        description:
          type: "string"
        adminClass:
          type: "boolean"
        active:
          type: "boolean"
    RoleFilter:
      type: "object"
      properties: { }
...